// src/utils/enums/campaignCommunityStatus.ts
export enum CampaignCommunityStatus {
  CampaignBriefSent = 'CampaignBriefSent',
  CampaignAccepted = 'CampaignAccepted',
  CampaignDeclined = 'CampaignDeclined',
  ContentApproved = 'ContentApproved',
  TaskCreated = 'TaskCreated',
  TaskRequestSent = 'TaskRequestSent',
  TaskScheduled = 'TaskScheduled',
  TaskDeclined = 'TaskDeclined',
  TaskCompleted = 'TaskCompleted',
  TaskPaused = 'TaskPaused',
  TaskFailed = 'TaskFailed',
  PaymentSent = 'PaymentSent',
}